requirements.txt
black==22.3.0
click==8.1.2
ImageHash==4.2.1
mypy-extensions==0.4.3
numpy==1.22.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
PyWavelets==1.3.0
scipy==1.8.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1